requirements.txt
Flask==3.0.3
Authlib==1.3.2
Jinja2==3.1.4
Werkzeug==3.0.3
itsdangerous==2.2.0

SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
xmltodict==0.13.0

google-cloud-storage==2.17.0
google-auth==2.33.0
boto3==1.34.162

gunicorn==22.0.0
cachetools==5.4.0
